+++ src/mesh.cpp
@@ -1,6 +1,6 @@
 #include "mesh.hpp"
 
-Mesh::Mesh(std::vector<Vertex> vertices, std::vector<unsigned short> indices, bool hasMX) {
+Mesh::Mesh(std::vector<Vertex> vertices, std::vector<GLuint> indices, bool hasMX) {
 	_vertices = vertices;
 	_indices = indices;
 	_hasModelMX = hasMX;
@@ -8,7 +8,7 @@ Mesh::Mesh(std::vector<Vertex> vertices, std::vector<unsigned short> indices, bo
 	_setupAttributes();
 }
 
-Mesh::Mesh(std::vector<Vertex> vertices, std::vector<unsigned short> indices, glm::vec3 min, glm::vec3 max, bool hasMX, std::vector<Triangle>& triangles) {
+Mesh::Mesh(std::vector<Vertex> vertices, std::vector<GLuint> indices, glm::vec3 min, glm::vec3 max, bool hasMX, std::vector<Triangle>& triangles) {
 	_vertices = vertices;
 	_indices = indices;
 	_min = min;
@@ -36,7 +36,7 @@ void Mesh::_setupBuffers() {
 
 	glGenBuffers(1, &_ibo);
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, _ibo);
-	glBufferData(GL_ELEMENT_ARRAY_BUFFER, _indices.size() * sizeof(unsigned short), &_indices[0], GL_STATIC_DRAW);
+	glBufferData(GL_ELEMENT_ARRAY_BUFFER, _indices.size() * sizeof(GLuint), &_indices[0], GL_STATIC_DRAW);
 }
 
 void Mesh::_setupAttributes() {
